+++ src/app/ChangeLogPreview.ts
@@ -3,14 +3,24 @@ import { ChangeLogPreviewInformation } from '../types/BaseTypes';
 const clp: ChangeLogPreviewInformation = {
   items: [
     {
-      title: 'Northwestern Discord Integration!',
+      title: 'The URL save data system is dead 💀',
       description:
-        'Easily share course information in chat while talking about classes with friends on the Northwestern Discord server! Just use "/course". Did you know you can see all of the latest dining hall menus in "#menu" without having to wait a millennium for Dine On Campus to load? And you can join thousands of other Northwestern students, alumni, applicants, faculty, and staff?? Join it now???? Or you\'ll be banned from Paper for life. Just kidding. But still.',
+        "The long URLs for plans and schedules that used to exist since the launch of the original Plan Northwestern are now gone. You can still share links to plans or schedules in the export menu. The shortened links are now the default. This change won't affect you if you just use an account or the browser storage to save your plans. Read more on my blog!",
+    },
+    {
+      title: 'Redesigned settings menu',
+      description:
+        'Enjoy your customization experience with a nicer settings menu.',
+    },
+    {
+      title: 'Duplicate plans and schedules',
+      description:
+        'Making multiple versions of the same base plan or schedule is now easier than ever. Hover over a plan or schedule saved to your account in the sidebar and press the duplicate button to create a new one with the same contents and notes.',
     },
   ],
   link: {
-    text: 'Join the NU Discord 🥺',
-    url: 'https://discord.gg/northwestern-655629737888055317',
+    text: 'What died?? Read on the blog.',
+    url: 'https://dilan.blog/discontinuing-the-paper-url-save-data-system/',
   },
 };